+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 
 ## Unreleased
 
+### Fixed
+- Remove redundant verbose error messages that were not different from messages that were already being printed.
+
 ### Changed
+- Use dedicated logstream failure category for param related error.
 - An authentication attempt with an expired auth token will result in a `auth token expired` error instead of `unauthorized`.
 - A successful authentication with an auth token will display a warning with time left before token expires if it's 14 days or under.
+ 
 
 ## v0.7.21 - 2023-10-24